+++ org-tweetyproject-causal/pom.xml
@@ -1,0 +1,30 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>causal</artifactId>
+
+    <name>TweetyProject (Causal Reasoning Library)</name>
+
+    <parent>
+        <groupId>org.tweetyproject</groupId>
+        <artifactId>parent-pom</artifactId>
+        <version>1.25-SNAPSHOT</version>
+        <relativePath>..</relativePath>
+    </parent>
+    <dependencies>
+        <dependency>
+            <groupId>org.tweetyproject.arg</groupId>
+            <artifactId>dung</artifactId>
+            <version>1.25-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.tweetyproject.arg</groupId>
+            <artifactId>deductive</artifactId>
+            <version>1.25-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
+</project>
+
