+++ org-tweetyproject-causal/src/main/java/org/tweetyproject/causal/reasoner/CausalReasoner.java
@@ -1,0 +1,35 @@
+/*
+ * This file is part of "TweetyProject", a collection of Java libraries for
+ * logical aspects of artificial intelligence and knowledge representation.
+ *
+ * TweetyProject is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Copyright 2024 The TweetyProject Team <http://tweetyproject.org/contact/>
+ */
+package org.tweetyproject.causal.reasoner;
+
+import org.tweetyproject.logics.pl.syntax.PlFormula;
+
+import java.util.Collection;
+
+public abstract class CausalReasoner {
+
+    protected Collection<PlFormula> observations;
+
+    /*
+    remove observations from CKB
+    CKB is a BeliefSet for the struct equations + a set of assumptions
+    observations only in the causal reasoner
+    causal reasoner has method to create AF, get conclusions, verify conclusions
+     */
+}
