+++ org-tweetyproject-causal/src/main/java/org/tweetyproject/causal/semantics/CounterfactualStatement.java
@@ -16,10 +16,10 @@
 *
 * Copyright 2024 The TweetyProject Team <http://tweetyproject.org/contact/>
 */
-package org.tweetyproject.arg.dung.causal.semantics;
+package org.tweetyproject.causal.semantics;
 
-import org.tweetyproject.arg.dung.causal.syntax.CausalKnowledgeBase;
-import org.tweetyproject.arg.dung.causal.syntax.StructuralCausalModel;
+import org.tweetyproject.causal.syntax.CausalKnowledgeBase;
+import org.tweetyproject.causal.syntax.StructuralCausalModel;
 import org.tweetyproject.logics.pl.syntax.PlFormula;
 import org.tweetyproject.logics.pl.syntax.Proposition;
