+++ org-tweetyproject-causal/src/main/java/org/tweetyproject/causal/semantics/InterventionalStatement.java
@@ -16,9 +16,9 @@
 *
 * Copyright 2024 The TweetyProject Team <http://tweetyproject.org/contact/>
 */
-package org.tweetyproject.arg.dung.causal.semantics;
+package org.tweetyproject.causal.semantics;
 
-import org.tweetyproject.arg.dung.causal.syntax.CausalKnowledgeBase;
+import org.tweetyproject.causal.syntax.CausalKnowledgeBase;
 import org.tweetyproject.logics.pl.syntax.PlFormula;
 import org.tweetyproject.logics.pl.syntax.Proposition;
