+++ org-tweetyproject-causal/src/main/java/org/tweetyproject/causal/syntax/CausalKnowledgeBase.java
@@ -16,7 +16,7 @@
  *
  * Copyright 2024 The TweetyProject Team <http://tweetyproject.org/contact/>
  */
-package org.tweetyproject.arg.dung.causal.syntax;
+package org.tweetyproject.causal.syntax;
 
 import org.tweetyproject.logics.pl.reasoner.SimplePlReasoner;
 import org.tweetyproject.logics.pl.syntax.Negation;
