+++ pom.xml
@@ -187,6 +187,7 @@
     <module>org-tweetyproject-action</module>
     <module>org-tweetyproject-beliefdynamics</module>
     <module>org-tweetyproject-cli</module>
+    <module>org-tweetyproject-causal</module>
     <module>org-tweetyproject-graphs</module>
     <module>org-tweetyproject-preferences</module>
     <module>org-tweetyproject-machinelearning</module>
